% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_avgSil.R
\name{check_avgSil}
\alias{check_avgSil}
\title{Check Average Silhouette score index}
\usage{
check_avgSil(data, range = 3:10)
}
\arguments{
\item{data}{Vector of abundance scores from a single sample.}

\item{range}{The range of values of k to test.}
}
\value{
Vector with average Silhouette score index for each pre-specified k.
}
\description{
Check Average Silhouette score index
}
\examples{
library(dplyr)
sample_2044662 <- nice_tidy \%>\% filter(Sample == "ERR2044662") \%>\% pull(Abundance)
check_CH(sample_2044662)

# To change range
check_CH(sample_2044662, range = 4:11)

}
