% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_CH.R
\name{check_CH}
\alias{check_CH}
\title{Check Calinsky-Harabasz index}
\usage{
check_CH(data, range = 3:10)
}
\arguments{
\item{data}{Vector of abundance scores from a single sample.}

\item{range}{The range of values of k to test.}
}
\value{
Vector with Calinsky-Harabasz index for each pre-specified k.
}
\description{
Check Calinsky-Harabasz index
}
\examples{
library(dplyr)
sample_2044662 <- nice_tidy \%>\% filter(Sample == "ERR2044662") \%>\% pull(Abundance)
check_CH(sample_2044662)

# To change range
check_CH(sample_2044662, range = 4:11)

}
