% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_DB_plot.R
\name{check_DB_plot}
\alias{check_DB_plot}
\title{Plots Davies-Bouldin Index for a single sample}
\usage{
check_DB_plot(data, range = 3:10)
}
\arguments{
\item{data}{Data frame with Davies-Bouldin scores for each k.}

\item{range}{The range of values of k to test, default is from 3 to 10.}
}
\value{
A plot with Davies-Bouldin score for each pre-specified k in a given sample.
}
\description{
Plots Davies-Bouldin Index for a single sample
}
\examples{

library(dplyr)
sample_2044662 <- nice_tidy \%>\% filter(Sample == "ERR2044662") \%>\% pull(Abundance)
check_DB(sample_2044662)

# To change range
check_DB(sample_2044662, range = 4:11)

}
