% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_rb.R
\name{define_rb}
\alias{define_rb}
\title{Define Rare Biosphere}
\usage{
define_rb(
  data,
  classification_vector = c("Rare", "Undetermined", "Abundant"),
  samples_id = "Sample",
  abundance_id = "Abundance"
)
}
\arguments{
\item{data}{A tibble with, at least, a column for Abundance and Sample. Additional columns are allowed.}

\item{classification_vector}{A vector of strings with the names for each cluster, from lower to higher abundance. Default is c("Rare", "Undetermined", "Abundance")}

\item{samples_id}{String with name of column with sample names}

\item{abundance_id}{String with name of column with abundance values}
}
\value{
A tibble with the initial columns and new columns for the cluster and classification of each species.
}
\description{
Define Rare Biosphere
}
\examples{
1

}
