% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_k.R
\name{evaluate_k}
\alias{evaluate_k}
\title{Evaluate k}
\usage{
evaluate_k(data, range = 3:10, ...)
}
\arguments{
\item{data}{Vector of abundance scores from a single sample.}

\item{range}{The range of values of k to test.}

\item{...}{Extra arguments.}
}
\value{
testing
}
\description{
Evaluate k
}
\examples{
#testing
}
