% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_sample_k.R
\name{evaluate_sample_k}
\alias{evaluate_sample_k}
\title{Evaluate sample k}
\usage{
evaluate_sample_k(data, range = 3:10, ...)
}
\arguments{
\item{data}{A tibble with, at least, a column for Abundance and Sample. Additional columns are allowed.}

\item{range}{The range of values of k to test, default is from 3 to 10.}

\item{...}{Extra arguments.}
}
\value{
testing
}
\description{
This functions calculates three indices (Davies-Bouldin, Calinsky-Harabasz and average Silhouette score) for each k.
Calculations are made for a single sample and for a default range of k that goes from 3 to 10.
}
\examples{
#testing
}
