% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nice_env}
\alias{nice_env}
\title{Metadata of samples from SSU rRNA amplicon of the Arctic Ocean during Winter-Spring Transition}
\format{
\subsection{\code{nice_tidy}}{

A data frame with 4716 rows and 10 columns:
\describe{
\item{Sample}{Sample ID used in the original study}
\item{ENA_ID}{Sample ID equivalent to Sample in the OTU table}
\item{Month}{Month of sampling}
\item{Region}{Ocean region of sampling event}
\item{Water.mass}{Water mass of sampling event}
\item{Latitude}{Latitude of sampling event}
\item{Longitude}{Longitude of sampling event}

}
}
}
\source{
\url{https://link.springer.com/article/10.1007/s00248-021-01821-9}
}
\usage{
nice_env
}
\description{
This dataset provides information on the samples used for rRNA amplicon of the Arctic Ocean during Winter-Spring Transition.
}
\keyword{datasets}
