% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nice_tidy}
\alias{nice_tidy}
\title{Tidy OTU table from SSU rRNA amplicon of the Arctic Ocean during Winter-Spring Transition}
\format{
## `nice_tidy`
A data frame with 4716 rows and 10 columns:
\describe{
  \item{Sample}{Sample ID}
  \item{Abundance}{Abundance}
  \item{OTU}{Taxonomic unit ID}
  \item{Domain}{Domain level classification of taxonomic unit}
  \item{Phylum}{Domain level classification of taxonomic unit}
  \item{Class}{Domain level classification of taxonomic unit}
  \item{Order}{Domain level classification of taxonomic unit}
  \item{Family}{Domain level classification of taxonomic unit}
  \item{Genus}{Domain level classification of taxonomic unit}
  \item{Species}{Domain level classification of taxonomic unit}
  ...
}
}
\source{
<https://www.ebi.ac.uk/metagenomics/studies/MGYS00001922#analysis>
}
\usage{
nice_tidy
}
\description{
Tidy version of original OTU table (nice).
A new column (Sample) includes the sample identifiers and a new column (Abundance) includes the abundance for each taxonomic unit.
For details on OTU table processing see help page of nice data.
}
\keyword{datasets}
