% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ulrb_clustering.R
\name{plot_ulrb_clustering}
\alias{plot_ulrb_clustering}
\title{Sanity check rare biosphere definition for one sample}
\usage{
plot_ulrb_clustering(
  data,
  sample_id = NULL,
  taxa_col,
  plot_all = FALSE,
  classification_col = "Classification",
  abundance_col = "Abundance",
  log_scaled = FALSE,
  colors = c("#009E73", "#F0E442", "#CC79A7"),
  ...
)
}
\arguments{
\item{data}{a data.frame with, at least, the classification, abundance and sample information for each taxonomic unit.}

\item{sample_id}{string with name of selected sample.}

\item{taxa_col}{string with name of column with taxonomic units. Usually OTU or ASV.}

\item{plot_all}{If TRUE, will make a plot for all samples with mean and standard deviation. If FALSE (default), then the plot will illustrate a single sample, that you have to specifiy in sample_id argument.}

\item{classification_col}{string with name of column with classification for each row. Default value is "Classification".}

\item{abundance_col}{string with name of column with abundance values. Default is "Abundance".}

\item{log_scaled}{if TRUE then abundance scores will be shown in Log10 scale. Default to FALSE.}

\item{colors}{vector with colors. Should have the same lenght as the number of classifications}

\item{...}{other arguments}
}
\value{
a ggplot object
}
\description{
Sanity check rare biosphere definition for one sample
}
\examples{
classified_species <- define_rb(nice_tidy)

plot_ulrb_clustering(classified_species,
                       sample_id = "ERR2044669",
                       taxa_col = "OTU",
                       abundance_col = "Abundance")

}
