% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ulrb_silhouette.R
\name{plot_ulrb_silhouette}
\alias{plot_ulrb_silhouette}
\title{Sanity check of Silhouette plot of selected sample}
\usage{
plot_ulrb_silhouette(
  data,
  sample_id = NULL,
  taxa_col,
  plot_all = FALSE,
  classification_col = "Classification",
  silhouette_score = "Silhouette_scores",
  colors = c("#009E73", "#F0E442", "#CC79A7"),
  log_scaled = FALSE,
  ...
)
}
\arguments{
\item{data}{...}

\item{sample_id}{string with name of selected sample.}

\item{taxa_col}{string with name of column with taxonomic units. Usually OTU or ASV.}

\item{plot_all}{If TRUE, will make a plot for all samples with mean and standard deviation. If FALSE (default), then the plot will illustrate a single sample, that you have to specifiy in sample_id argument.}

\item{classification_col}{string with name of column with classification for each row. Default value is "Classification".}

\item{silhouette_score}{string with column name with silhouette score values. Default is "Silhouette_scores"}

\item{colors}{vector with colors. Should have the same lenght as the number of classifications.}

\item{log_scaled}{if TRUE then abundance scores will be shown in Log10 scale. Default to FALSE.}

\item{...}{other arguments.}
}
\value{
A ggplot object of Silhouette plot obtained from the selected sample.
}
\description{
Sanity check of Silhouette plot of selected sample
}
\examples{
plot_ulrb_silhouette(data = define_rb(nice_tidy, simplified = FALSE),
 taxa_col = "OTU", sample_id = "ERR2044662")

}
