% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_check_all.R
\name{sanity_check_all}
\alias{sanity_check_all}
\title{Sanity check of results for all samples}
\usage{
sanity_check_all(data, ...)
}
\arguments{
\item{data}{output from sanity_check_sample_rb}

\item{...}{other arguments}
}
\value{
1
}
\description{
Sanity check of results for all samples
}
\examples{
1
}
