% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_check_sample_rb.R
\name{sanity_check_sample_rb}
\alias{sanity_check_sample_rb}
\title{Sanity check rare biosphere definition for one sample}
\usage{
sanity_check_sample_rb(
  data,
  sample_id,
  taxa_id,
  classification_id = "Classification",
  abundance_id = "Abundance",
  colors = c("#0072B2", "#D55E00", "#CC79A7"),
  log_scaled = FALSE,
  ...
)
}
\arguments{
\item{data}{a data.frame with, at least, the classification, abundance and sample information for each taxonomic unit.}

\item{sample_id}{string with name of selected sample.}

\item{taxa_id}{string with name of column with taxonomic units. Usually OTU or ASV.}

\item{classification_id}{string with name of column with classification for each row. Default value is "Classification".}

\item{abundance_id}{string with name of column with abundance values. Default is "Abundance".}

\item{colors}{vector with colors. Should have the same lenght as the number of classifications}

\item{log_scaled}{if TRUE then abundance scores will be shown in Log10 scale. Default to FALSE.}

\item{...}{other arguments}
}
\value{
a ggplot object
}
\description{
Sanity check rare biosphere definition for one sample
}
\examples{
classified_species <- define_rb(nice_tidy)

sanity_check_sample_rb(classified_species,
                       sample_id = "ERR2044669",
                       taxa_id = "OTU",
                       abundance_id = "Abundance")

}
