% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_check_sample_sil.R
\name{sanity_check_sample_sil}
\alias{sanity_check_sample_sil}
\title{Sanity check of Silhouette plot of selected sample}
\usage{
sanity_check_sample_sil(data)
}
\arguments{
\item{data}{...}
}
\value{
A ggplot object of Silhouette plot obtained from the selected sample.
}
\description{
Sanity check of Silhouette plot of selected sample
}
\examples{
sanity_check_sample_sil(data = define_rb(nice_tidy, simplified = FALSE))
}
