% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulrb-package.R
\docType{package}
\name{ulrb-package}
\alias{ulrb}
\alias{ulrb-package}
\title{ulrb: Unsupervised Learning Based Definition Of Microbial Rare Biosphere}
\description{
The ulrb package solves the problem of the definition of rarity in molecular microbiology by replacing human decision with an unsupervised machine learning algorithm (partitionin around medoids, or k-medoids). This algorithm works for any type of microbiome data, provided there is an abundance score for each taxonomic unit. For validation of this method to several kinds of molecular data and environments, please see Pascoal et al, 2023.
}
\author{
\strong{Maintainer}: Francisco Pascoal \email{fpascoal1996@gmail.com}

Authors:
\itemize{
  \item Paula Branco \email{paobranco@gmail.com}
  \item Luís Torgo \email{ltorgo@dal.ca}
  \item Rodrigo Costa \email{rodrigoscosta@tecnico.ulisboa.pt}
  \item Catarina Magalhães \email{catarinamagalhaes1972@gmail.com}
}

}
\keyword{internal}
