% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_DB.R
\name{check_DB}
\alias{check_DB}
\title{Check Davies-Bouldin Index}
\usage{
check_DB(
  data,
  sample_id,
  samples_col = "Sample",
  abundance_col = "Abundance",
  range = 3:10,
  with_plot = FALSE,
  ...
)
}
\arguments{
\item{data}{A tibble with, at least, a column for Abundance and Sample. Additional columns are allowed.}

\item{samples_col}{String with name of column with sample names}

\item{abundance_col}{String with name of column with abundance values}

\item{range}{The range of values of k to test, default is from 3 to 10.}

\item{with_plot}{If FALSE (default) returns a vector, but if TRUE will returna plot with the scores.}

\item{...}{Extra arguments.}
}
\value{
A vector or plot with Davies-Bouldin index for each pre-specified k in a given sample.
}
\description{
Check Davies-Bouldin Index
}
\examples{

library(dplyr)
sample_2044662 <- nice_tidy \%>\% filter(Sample == "ERR2044662") \%>\% pull(Abundance)
check_DB(sample_2044662)

# To change range
check_DB(sample_2044662, range = 4:11)

}
