% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_DB.R
\name{check_DB}
\alias{check_DB}
\title{Check Davies-Bouldin Index}
\usage{
check_DB(data, range = 3:10)
}
\arguments{
\item{data}{Vector of abundance scores from a single sample.}

\item{range}{The range of values of k to test.}
}
\description{
Check Davies-Bouldin Index
}
\examples{

library(dplyr)
sample_2044662 <- nice_tidy \%>\% filter(Sample == "ERR2044662") \%>\% pull(Abundance)
check_DB(sample_2044662)

# To change range
check_DB(sample_2044662, range = 4:11)

}
