% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_rb.R
\name{define_rb}
\alias{define_rb}
\title{Define Rare Biosphere}
\usage{
define_rb(
  data,
  classification_vector = c("Rare", "Undetermined", "Abundant"),
  samples_col = "Sample",
  abundance_col = "Abundance",
  simplified = FALSE,
  automatic = FALSE,
  index = "Average Silhouette Score",
  ...
)
}
\arguments{
\item{data}{A tibble with, at least, a column for Abundance and Sample. Additional columns are allowed.}

\item{classification_vector}{A vector of strings with the names for each cluster, from lower to higher abundance. Default is c("Rare", "Undetermined", "Abundance").}

\item{samples_col}{String with name of column with sample names.}

\item{abundance_col}{String with name of column with abundance values.}

\item{simplified}{Can be TRUE/FALSE. Default (FALSE) provides an additional column with detailed pam() results
and Silhouette scores. If TRUE, only the Classification result is added to the original input data.}

\item{automatic}{By default (FALSE), will assume a classification into "Rare", "Undetermined" or "Abundant". If TRUE, then it will automatically select the number of classifications (or k),
based on the index argument.}

\item{index}{Index used to select best k. Can be one of: "Average Silhouette Score", "Davies-Bouldin" or "Calinsky-Harabasz".}

\item{...}{Extra arguments.}
}
\value{
The input data.frame with extra columns containing the classification ()and additional metrics (if detailed = TRUE).
}
\description{
Classifies the species in each sample into either "Rare", "Undetermined" or "Abundant".
}
\details{
\strong{Overview}

Function to cluster species abundance with k-medoid algorithm. By default, We propose the division into three clusters (k = 3),
which can have the convenient description of: "rare", "undetermined" and "abundant". The species from the cluster with lowest median abundance
is considered to be the "rare biosphere".

The classification vector

The classification vector (argument classification_vector) represents the different clusters to be used, by ascending order of median abundance.
To change the number of clusters, change the number of elements in the classification vector, but order matters! Depending on the number of clusters used, you can change the meaning that best applies to your research. For example,
you can use a classification vector with the designations: "very rare", "rare", "abundant" and "very abundant"; which would apply a k = 4 underneath.
It is possible to use any number of clusters, as long as they are within 2 and the maximum possible k. The maximum possible k is the
number of different abundance scores observed in a single sample. Note, however, that we do not recommend any clustering for k > 10
and we also don't recommend k = 2 (we explain in more detail in Pascoal et al., 2023; and in the "explore-classifications".

Automatic selection of cluster numbers

To automatically decide the number of clusters (i.e., the value of k), it is possible to do so with the argument @param \code{automatic=TRUE}. For details on
complete automation of define_rb(), please see the documentation for suggest_k(). Briefly, the k with best average Silhouette score
is selected, from a range of k values between 3 and 10. It is possible to decide k based on other indices ("Davies-Bouldin" or "Calinsky-Harabasz").

If you want a more fine grained analysis of k values, we provide other functions to help you, see \strong{missing vignette}.

Verify clustering results

If half of the taxa of any cluster got a Silhouette score below 0.5 in any sample, then a warning is provided. The warning provides the number of times this issue occurred.
You can inspect other alternatives to reduce the number of bad clusterings, but it is possible that, in some situations, you just can't find an optimal clustering.

The detailed output gives you access to all of the clustering results:
\itemize{
\item "pam_object" is a list with the original results from the k-medoids clustering, see pam() documentation.
\item "Level" is an integer indicating the specific cluster attributed by the pam() function for each observation. Its order is random.
\item "Silhouette_scores" provides the Silhouette score obtained for each observation, i.e. a score for each taxa.
\item "Cluster_median_abundance" provides the median taxa abundance of each cluster.
\item "median_Silhouette" provides the median Silhouette score obtained for each cluster.
\item "Evaluation" indicates if the silhouette score obtained for a given observation is below the median Silhouette of its cluster and sample.
}

You can make your own plots and analysis, but we also provide another function, plot_ulrb(), which illustrates the results obtained.

Notes

Understand that ulrb considers each sample as an independent community of taxa, which means clustering is also independent across different samples.
Thus, be aware that you will have clustering results and metrics for each single sample, which is why we also provide some functions to analyze results across
any number of samples (see: \link{plot_ulrb} for clustering results and \code{evaluate_k} for k selection).
}
\examples{

# Sample ID's
sample_names <- c("ERR2044662", "ERR2044663", "ERR2044664",
                   "ERR2044665", "ERR2044666", "ERR2044667",
                   "ERR2044668", "ERR2044669", "ERR2044670")

# If data is in wide format, with samples in cols
nice_tidy <- prepare_tidy_data(nice, sample_names = sample_names, samples_in = "cols")

# If table is in tidy format with a column for Samples and another for abundance
define_rb(nice_tidy)

# Automatic decision, instead of a predefined definition
define_rb(nice_tidy, automatic = TRUE)

# Automatic decision, using Davies-Boulding index, instead of average Silhouette score (default)
define_rb(nice_tidy, automatic = TRUE, index = "Davies-Bouldin")

# User defined classifications
user_classifications <- c("very rare", "rare", "undetermined", "abundant", "very abundant")

define_rb(nice_tidy, classification_vector = user_classifications)

# Easy to incorporate in big pipes
nice_tidy \%>\%
 filter(Domain != "sk__Archaea") \%>\% # Removed Archaea
 filter(Abundance > 10) \%>\% #Removed taxa below 10 reads
 define_rb(classification_vector = c("very rare", "rare", "abundant", "very abundant")) # Classify according to a different set of classifications

}
\seealso{
\code{\link[=suggest_k]{suggest_k()}}, \code{\link[=evaluate_k]{evaluate_k()}}, \code{\link[=plot_ulrb]{plot_ulrb()}}, \code{\link[cluster:pam]{cluster::pam()}}
}
