% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_rb.R
\name{define_rb}
\alias{define_rb}
\title{Define Rare Biosphere}
\usage{
define_rb(
  data,
  classification_vector = c("Rare", "Undetermined", "Abundant"),
  samples_col = "Sample",
  abundance_col = "Abundance",
  simplified = FALSE,
  automatic = FALSE,
  index = "Average Silhouette Score",
  ...
)
}
\arguments{
\item{data}{A tibble with, at least, a column for Abundance and Sample. Additional columns are allowed.}

\item{classification_vector}{A vector of strings with the names for each cluster, from lower to higher abundance. Default is c("Rare", "Undetermined", "Abundance")}

\item{samples_col}{String with name of column with sample names}

\item{abundance_col}{String with name of column with abundance values}

\item{simplified}{Can be TRUE/FALSE. Default (FALSE) provides an additional column with detailed pam() results (pam_details) and Silhouette scores (Silhouette_scores).}

\item{automatic}{If TRUE, then it will automatically select the number of classifications (or k),
based on the index argument.}

\item{index}{Index used to select best k. Can be one of: "Average Silhouette Score", "Davies-Bouldin" or "Calinsky-Harabasz".}

\item{...}{Extra arguments.}
}
\value{
A tibble with the initial columns and new columns for the cluster and classification of each species.
}
\description{
Define Rare Biosphere
}
\examples{

define_rb(nice_tidy)

# If data is in wide format, with samples in cols
sample_names <- c("ERR2044662", "ERR2044663", "ERR2044664",
                   "ERR2044665", "ERR2044666", "ERR2044667",
                   "ERR2044668", "ERR2044669", "ERR2044670")

nice_tidy <- prepare_tidy_data(nice, sample_names = sample_names, samples_in = "cols")
define_rb(nice_tidy)

# If data is in wide format, with samples in rows
### do this later ###

}
