% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_k.R
\name{evaluate_k}
\alias{evaluate_k}
\title{Evaluate k from all samples in a dataset}
\usage{
evaluate_k(
  data,
  range = 3:10,
  samples_col = "Sample",
  abundance_col = "Abundance",
  with_plot = FALSE,
  ...
)
}
\arguments{
\item{data}{A tibble with, at least, a column for Abundance and Sample. Additional columns are allowed.}

\item{range}{The range of values of k to test, default is from 3 to 10.}

\item{samples_col}{String with name of column with sample names.}

\item{abundance_col}{String with name of column with abundance values.}

\item{with_plot}{If FALSE (default) returns a vector, but if TRUE will return a plot with the scores.}

\item{...}{Extra arguments.}
}
\value{
Tidy table with three indices for each k within each sample.
}
\description{
Evaluate k from all samples in a dataset
}
\examples{
evaluate_k(nice_tidy)


}
