% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_sample_k.R
\name{evaluate_sample_k}
\alias{evaluate_sample_k}
\title{Evaluate sample k}
\usage{
evaluate_sample_k(data, range = 3:10, ...)
}
\arguments{
\item{data}{Vector of abundance scores from a single sample.}

\item{range}{The range of values of k to test.}

\item{...}{Extra arguments.}
}
\value{
testing
}
\description{
This functions calculates three indices (Davies-Bouldin, Calinsky-Harabasz and average Silhouette score) for each k.
Calculations are made for a single sample.
}
\examples{
#testing
}
