% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nice_tidy}
\alias{nice_tidy}
\title{SSU rRNA amplicon of the Arctic Ocean during Winter-Spring Transition}
\format{
## `nice_tidy`
A data frame with 524 rows and 17 columns:
\describe{
  \item{ERR2044662, "ERR2044663", "ERR2044664","ERR2044665", "ERR2044666", "ERR2044667","ERR2044668", "ERR2044669", "ERR2044670"}{Country name}
  \item{OTU}{Taxonomic unit ID}
  \item{Domain}{Domain level classification of taxonomic unit}
  \item{Phylum}{Domain level classification of taxonomic unit}
  \item{Class}{Domain level classification of taxonomic unit}
  \item{Order}{Domain level classification of taxonomic unit}
  \item{Family}{Domain level classification of taxonomic unit}
  \item{Genus}{Domain level classification of taxonomic unit}
  \item{Species}{Domain level classification of taxonomic unit}
  ...
}
}
\source{
<https://www.ebi.ac.uk/metagenomics/api/v1/studies/MGYS00001922/pipelines/5.0/file/ERP024265_taxonomy_abundances_SSU_v5.0.tsv>
}
\usage{
nice_tidy
}
\description{
A raw OTU table from SSU rRNA amplicon of the Arctic Ocean during Winter-Spring Transition.
Raw reads were processed into OTUs by the MGnify platform (v5). Original .fastq files with
sequences are publicly available in European Nucleotide Archive under accesion number XXXXXX
}
\keyword{datasets}
