% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ulrb.R
\name{plot_ulrb}
\alias{plot_ulrb}
\title{Plot relevant metrics obtained from ulrb clustering}
\usage{
plot_ulrb(
  data,
  see = "clustering",
  sample_id = NULL,
  taxa_col,
  classification_col = "Classification",
  abundance_col = "Abundance",
  colors = c("#0072B2", "#D55E00", "#CC79A7"),
  log_scaled = FALSE,
  ...
)
}
\arguments{
\item{data}{a data.frame with, at least, the classification, abundance and sample information for each taxonomic unit.}

\item{see}{What you want to see in the plot. Can be either "clustering" (default), "Silhouette score", "Calinsky-Harabasz" or "Davies-Bouldin".}

\item{sample_id}{Default is NULL. If you want to analyze one specific sample, specify its name as a string.}

\item{taxa_col}{string with name of column with taxonomic units, e.g., OTU, ASV...}

\item{classification_col}{string with name of column with classification for each row. Default value is "Classification".}

\item{abundance_col}{string with name of column with abundance values. Default is "Abundance".}

\item{colors}{vector with colors. Should have the same length as the number of classifications}

\item{log_scaled}{if TRUE then abundance scores will be shown in Log10 scale. Default to FALSE.}

\item{...}{other arguments}
}
\value{
a ggplot object
}
\description{
Plot relevant metrics obtained from ulrb clustering
}
\examples{
classified_species <- define_rb(nice_tidy)

plot_ulrb(classified_species,
                       sample_id = "ERR2044669",
                       taxa_col = "OTU",
                       abundance_col = "Abundance")

}
