% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ulrb_silhouette.R
\name{plot_ulrb_silhouette}
\alias{plot_ulrb_silhouette}
\title{Sanity check of Silhouette plot of selected sample}
\usage{
plot_ulrb_silhouette(
  data,
  sample_id,
  taxa_col,
  classification_col = "Classification",
  silhouette_score = "Silhouette_scores",
  colors = c("#0072B2", "#D55E00", "#CC79A7"),
  ...
)
}
\arguments{
\item{data}{...}

\item{sample_id}{string with name of selected sample.}

\item{taxa_col}{string with name of column with taxonomic units. Usually OTU or ASV.}

\item{classification_col}{string with name of column with classification for each row. Default value is "Classification".}

\item{silhouette_score}{string with column name with silhouette score values. Default is "Silhouette_scores"}

\item{colors}{vector with colors. Should have the same lenght as the number of classifications}

\item{...}{other arguments}
}
\value{
A ggplot object of Silhouette plot obtained from the selected sample.
}
\description{
Sanity check of Silhouette plot of selected sample
}
\examples{
plot_ulrb_silhouette(data = define_rb(nice_tidy, simplified = FALSE),
 taxa_col = "OTU", sample_id = "ERR2044662")

}
