% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_tidy_data.R
\name{prepare_tidy_data}
\alias{prepare_tidy_data}
\title{Prepare data in tidy format}
\usage{
prepare_tidy_data(data, sample_names, samples_in = "cols", ...)
}
\arguments{
\item{data}{a data.frame in "wide" format, with samples in either columns or rows. This data.frame should not include any data besides abundance values per sample, per taxonomic unit. Additional data (e.g. taxonomy details) should be added afterwards.}

\item{sample_names}{a vector with the name of all samples.}

\item{samples_in}{a vector specifyng the location of the samples. It can either be "cols" (default) if samples are in columns, or "rows" if samples are in rows.}

\item{...}{additional arguments}
}
\value{
a tibble with with the arguments for define_rb() in tidy format.
}
\description{
Prepare data in tidy format
}
\examples{
1

}
