% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_check_all.R
\name{sanity_check_all}
\alias{sanity_check_all}
\title{Sanity check of results for all samples}
\usage{
sanity_check_all(
  data,
  sample_names,
  taxa_col,
  classification_col = "Classification",
  abundance_col = "Abundance",
  silhouette_score = "Silhouette_scores",
  colors = c("#0072B2", "#D55E00", "#CC79A7"),
  export_output = FALSE,
  output_name = "Sanity check figures",
  log_scaled = FALSE,
  ...
)
}
\arguments{
\item{data}{a tidy data.frame with samples, abundance and classification for each taxonomic unit.}

\item{sample_names}{a vector with the name of all samples.}

\item{taxa_col}{string with name of column with taxonomic units. Usually OTU or ASV.}

\item{classification_col}{string with name of column with classification for each row. Default value is "Classification".}

\item{abundance_col}{string with name of column with abundance values. Default is "Abundance".}

\item{silhouette_score}{string with column name with silhouette score values. Default is "Silhouette_scores"}

\item{colors}{vector with colors. Should have the same lenght as the number of classifications}

\item{export_output}{can be "yes" (default) or "no". If "yes", then a pdf file with all plots is produced,
else the plots are not printed into pdf and a are returned to R session instead.}

\item{output_name}{name of pdf file with ouput. Only works if export_output = "yes".}

\item{log_scaled}{if TRUE then abundance scores will be shown in Log10 scale. Default to FALSE.}

\item{...}{other arguments}
}
\value{
1
}
\description{
Sanity check of results for all samples
}
\examples{
#
sample_names <- c("ERR2044662", "ERR2044663", "ERR2044664",
                   "ERR2044665", "ERR2044666", "ERR2044667",
                   "ERR2044668", "ERR2044669", "ERR2044670")
#
classified_species <- define_rb(nice_tidy)
#
sanity_check_all(classified_species,
                 sample_names = sample_names,
                 taxa_col = "OTU",
                 abundance_col = "Abundance",
                 export_output = "no",
                 classification_col = "Classification")

}
