% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_k.R
\name{suggest_k}
\alias{suggest_k}
\title{Suggest k}
\usage{
suggest_k(
  data,
  range = 3:10,
  samples_col = "Sample",
  abundance_col = "Abundance",
  index = "Average Silhouette Score",
  detailed = FALSE,
  ...
)
}
\arguments{
\item{data}{a data.frame with, at least, the classification, abundance and sample information for each taxonomic unit.}

\item{range}{The range of values of k to test, default is from 3 to 10.}

\item{abundance_col}{string with name of column with abundance values. Default is "Abundance".}

\item{index}{Index used to select best k. Can be one of: "Average Silhouette Score", "Davies-Bouldin" or "Calinsky-Harabasz".}

\item{detailed}{If False (default) returns an integer with best overall k. If TRUE, returns a list with full details.}

\item{...}{Extra arguments.}
}
\value{
Integer indicating best k from selected index. Optionally, can return a list with details.
}
\description{
The best k is selected for each sample, based on the selected index. If different k's are obtained for different samples
(probable) then we calculate the mean value of k and return it as an integer.
}
\examples{
suggest_k(nice_tidy, detailed = TRUE)

suggest_k(nice_tidy, detailed = FALSE)

}
